refactor: 💡 protobuf nomenclature and optimisation
All protobufs with payloads distinguish now between ABI encoded payloads
meant to be fed directly to smart contracts, and otherwise arbitrary
bytes defined in implementation specific deployments. Timestamps were
optimised for handling, as well as wads / numbers that are greater than
64 bit are by default now strings for the purposes of protobuf. BidItem
was removed as this was just a wrapper for `bytes32`, therefore
`bytes32` is used in place now (as `bytes` in protobuf).

BREAKING CHANGE: 🧨 yes