fix: 🐛 eip712 storage signing
By removing the `SignedPayloadWrapper` we can make use of the same
signing routines that are used for the EIP-712 messages that are
exchanged via waku. In doing so, we introduce a `signature` field in the
`ServiceProviderData` message, and then we can use type omissions to
pass in unsigned data for signign and returned signed etc, same as we do
for messages.