fix: 🐛 standardize throws
This fix now standardizes throws and boolean returns such that if a
boolean function is called, and subsequently internally this throws an
exception, the exception is eaten and false is returned. When doing
general signing, if the EIP-712 signature fails, this throw is passed to
the calling function to be dealt with.

✅ Closes: 32